V15.4 SP5: GPU-native batched RCC search + engine perf pass
v15_gpu_search: batched enumeration (intercept-aimed), stage-1/stage-2 combo
rollouts, pick — all tensor ops, no V7.
v15_gpu_sim: vectorised combat (scatter_add, no N*P Python loop), dtype-
flexible (float32 for speed; float64 path still bit-exact, 120/120), launch
loop skipped on passive rollout steps.

Perf so far: 31h/5000-game gen — kernel-launch overhead dominates; needs
fusion (torch.compile) to beat the CPU path.

+++ tests/test_gpu_sim_equivalence.py
@@ -112,7 +112,7 @@ def run(n_players, n_states, rng):
     # CPU reference
     cpu_next = [fsim.step(s, a) for s, a in zip(states, actions)]
     # GPU batched
-    batch = gsim.from_faststates(states, device="cpu")
+    batch = gsim.from_faststates(states, device="cpu", dtype=torch.float64)
     at = _actions_tensor(actions, n_players)
     gpu_batch = gsim.step(batch, at)
     gpu_next = [gsim.to_faststate(gpu_batch, b) for b in range(len(states))]
